alterações nos desafios

+++ desafios/d007.py
@@ -1,5 +1,5 @@
-n1 = int(input('Digite a nota 1: '))
-n2 = int(input('Digite a nota 2: '))
-n3 = int(input('Digite a nota 3: '))
+n1 = float(input('Digite a nota 1: '))
+n2 = float(input('Digite a nota 2: '))
+n3 = float(input('Digite a nota 3: '))
 m = (n1+n2+n3)/3
 print('A nota média é {}'.format(m))
